Address PR review
