+++ server/resolvers/TestPlanRunOperations/deleteTestResultsResolver.js
@@ -3,10 +3,7 @@ const {
     updateTestPlanRun
 } = require('../../models/services/TestPlanRunService');
 const populateData = require('../../services/PopulatedData/populateData');
-const conflictsResolver = require('../TestPlanReport/conflictsResolver');
-const {
-    updateTestPlanReport
-} = require('../../models/services/TestPlanReportService');
+const persistConflictsCount = require('../helpers/persistConflictsCount');
 
 const deleteTestResultsResolver = async (
     { parentContext: { id: testPlanRunId } },
@@ -27,22 +24,10 @@ const deleteTestResultsResolver = async (
 
     await updateTestPlanRun(testPlanRunId, { testResults: [] });
 
-    // Perform in background
+    // TODO: Avoid blocking loads in test runs with a larger amount of tests
+    //       and/or test results
     await persistConflictsCount(testPlanRun);
     return populateData({ testPlanRunId });
-};
-
-const persistConflictsCount = async testPlanRun => {
-    const { testPlanReport: updatedTestPlanReport } = await populateData({
-        testPlanRunId: testPlanRun.id
-    });
-    const conflicts = await conflictsResolver(updatedTestPlanReport);
-    await updateTestPlanReport(updatedTestPlanReport.id, {
-        metrics: {
-            ...updatedTestPlanReport.metrics,
-            conflictsCount: conflicts.length
-        }
-    });
 };
 
 module.exports = deleteTestResultsResolver;
