+++ server/resolvers/TestResultOperations/saveTestResultCommon.js
@@ -7,10 +7,7 @@ const deepCustomMerge = require('../../util/deepCustomMerge');
 const deepPickEqual = require('../../util/deepPickEqual');
 const convertTestResultToInput = require('../TestPlanRunOperations/convertTestResultToInput');
 const createTestResultSkeleton = require('../TestPlanRunOperations/createTestResultSkeleton');
-const {
-    updateTestPlanReport
-} = require('../../models/services/TestPlanReportService');
-const conflictsResolver = require('../TestPlanReport/conflictsResolver');
+const persistConflictsCount = require('../helpers/persistConflictsCount');
 
 const saveTestResultCommon = async ({
     testResultId,
@@ -84,19 +81,6 @@ const saveTestResultCommon = async ({
     //       and/or test results
     await persistConflictsCount(testPlanRun);
     return populateData({ testResultId });
-};
-
-const persistConflictsCount = async testPlanRun => {
-    const { testPlanReport: updatedTestPlanReport } = await populateData({
-        testPlanRunId: testPlanRun.id
-    });
-    const conflicts = await conflictsResolver(updatedTestPlanReport);
-    await updateTestPlanReport(updatedTestPlanReport.id, {
-        metrics: {
-            ...updatedTestPlanReport.metrics,
-            conflictsCount: conflicts.length
-        }
-    });
 };
 
 const assertTestResultIsValid = newTestResult => {
